chore: use new build image
* update jenkinsfile to pipeline syntax

in order to build with java 17, quite a few changes needed to be done:

* upgrade gradlew to a recent version
* modernise build.gradle (especially the publishing steps)
* upgrade spotbugs
* fix a javadoc error
* update spring framework/boot versions (see below)
* update junit and mockito versions
* fix usage of now-removed `EnvironmentTestUtils` method

about spring versions:

spring framework:

rationale: use 6.0.x for compile, 5.3.x for test, since the latter
supports java 11

https://github.com/spring-projects/spring-framework/wiki/Spring-Framework-Versions

"6.0.x is the main production line as of November 2022. This new
generation of the framework comes with a JDK 17 and Jakarta EE 9
baseline.  5.3.x is the final feature branch of the 5th generation,
with long-term support provided on JDK 8, JDK 11, JDK 17 and the Java
EE 8 level."

spring boot:

rationale: use 3.0.x for compile, 2.1.x for test, since the latter
supports java 11

https://docs.spring.io/spring-boot/docs/3.0.4/reference/html/getting-started.html

"Spring Boot 3.0.4 requires Java 17 and is compatible up to and
including Java 19. Spring Framework 6.0.6 or above is also required."

https://docs.spring.io/spring-boot/docs/2.1.16.RELEASE/reference/html/getting-started-system-requirements.html

"Spring Boot 2.1.16.RELEASE requires Java 8 and is compatible up to
Java 12 (included)."